refactor: decouple rpc transport from git
