+++ src-tauri/src/instructions/rpc_protocol.md
@@ -1,16 +1,39 @@
-## Quorum app actions (file RPC)
+## Quorum app RPC
 
-Your writes are confined to your worktree. When a task needs an action the app must run for you, ask for it through the file mailbox at `$QUORUM_RPC_DIR` and wait for the reply. Only the ops listed below are accepted; anything else is rejected.
+Your writes are confined to your worktree. When an agent needs the app to do
+something it cannot do itself, send a JSON request through the mailbox at
+`$QUORUM_RPC_DIR` and wait for the reply.
 
-To make a call:
+Protocol shape:
 
-1. Choose a unique id (a UUID works).
-2. Write the request **atomically**: write `$QUORUM_RPC_DIR/requests/<id>.json.tmp`, then rename it to `$QUORUM_RPC_DIR/requests/<id>.json`. Body: `{"id":"<id>","op":"<op>","args":{}}`.
-3. Poll for `$QUORUM_RPC_DIR/responses/<id>.json` until it appears (every ~0.2s, with your own timeout), then read it.
+1. Pick a unique id, usually a UUID.
+2. Write the request atomically to `$QUORUM_RPC_DIR/requests/<id>.json.tmp`,
+   then rename it to `$QUORUM_RPC_DIR/requests/<id>.json`.
+3. Poll `$QUORUM_RPC_DIR/responses/<id>.json` until it appears, then read it.
 
-A success response is `{"id","ok":true,"exit_code","stdout","stderr"}`; a failure is `{"id","ok":false,"error":"..."}`.
+Request body:
 
-Wait for a reply (shell):
+```json
+{"id":"<id>","op":"<op>","args":{}}
+```
+
+`op` is an application-defined message name. `args` is an arbitrary JSON object
+for that operation's payload. The app replies with one final response payload;
+there is no streaming.
+
+A success response is:
+
+```json
+{"id":"<id>","ok":true,"exit_code":0,"stdout":"","stderr":""}
+```
+
+A failure response is:
+
+```json
+{"id":"<id>","ok":false,"error":"..."}
+```
+
+Example:
 
 ```sh
 ID=$(uuidgen)
@@ -20,28 +43,20 @@ until [ -f "$QUORUM_RPC_DIR/responses/$ID.json" ]; do sleep 0.2; done
 cat "$QUORUM_RPC_DIR/responses/$ID.json"
 ```
 
-For args carrying free text (commit messages, PR bodies), do NOT hand-escape the string into a `printf`/`echo` format — `\n` and quotes get mangled by the shell. Build the body with `jq -n --arg` from a heredoc, which escapes everything correctly on the first try:
+For free-text args, build the JSON with `jq -n --arg` so quotes and newlines
+are escaped correctly:
 
 ```sh
 ID=$(uuidgen)
 MSG=$(cat <<'EOF'
-feat: add the thing
-
-Optional longer body, safely multi-line.
+hello from the agent
 EOF
 )
-jq -n --arg id "$ID" --arg msg "$MSG" '{id:$id,op:"git_commit",args:{message:$msg}}' \
+jq -n --arg id "$ID" --arg msg "$MSG" '{id:$id,op:"echo",args:{message:$msg}}' \
   > "$QUORUM_RPC_DIR/requests/$ID.json.tmp"
 mv "$QUORUM_RPC_DIR/requests/$ID.json.tmp" "$QUORUM_RPC_DIR/requests/$ID.json"
 ```
 
-Available ops:
-
-- `ping` — liveness check; returns `pong`. No args.
-- `git_status` — runs `git status` in your worktree and returns its output. No args.
-- `git_commit` — stages all changes in your worktree and commits them. `args.message` (required) is the commit message.
-- `open_pr` — creates your branch (if you don't have one yet), pushes it, and opens a pull request against your base branch. `args.branch` is the branch name to create — choose a conventional, descriptive name like `fix/login-crash`, `feat/csv-export`, or `chore/bump-deps` (no `quorum/` prefix). Required the first time you push; omit it later to keep using your current branch (pushing again updates the same PR). Pass a *new* `args.branch` to cut a fresh branch for a second PR from the same worktree. `args.title` and `args.body` set the PR title and description; omit `title` to auto-fill from your commits.
-- `git_push` — pushes your branch to origin (sets the upstream on first push). When you're already on a branch, no args are needed — use it to update an existing pull request after committing fixes. When you don't have a branch yet, pass `args.branch` (a conventional name like `fix/…`) so it can be created at your current commit before pushing.
-- `git_update_branch` — fetches your base branch from origin and merges it into your current branch. No args. A non-zero `exit_code` with `CONFLICT` in `stdout` means the merge stopped on conflicts: resolve the listed files in your worktree, then complete the merge with `git_commit` and push with `git_push`.
-
-Use these ops for git: your worktree's git database lives outside the sandbox, so running `git add`/`commit`/`merge`/`push` yourself fails with a permission error. Commit through `git_commit`, push through `git_push`, sync your base branch through `git_update_branch`, and open PRs through `open_pr` instead.
+The exact op names are feature-specific. Quorum currently uses this same
+transport for Git actions, but future app features can define their own ops on
+top of the same mailbox.
